+++ lib/archive_validator/others_relayables_validator.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class ArchiveValidator
+  class OthersRelayablesValidator < CollectionValidator
+    def collection
+      others_relayables
+    end
+
+    def entity_validator
+      RelayableValidator
+    end
+  end
+end